requirements.txt
firebase-admin>=6.2.0
folium>=0.14.0
geopy>=2.3.0
numpy>=1.24.0
streamlit>=1.26.0
streamlit-folium>=0.13.0
rdp>=0.8.0
streamlit_authenticator>=0.2.2
selenium>=4.11.0
pandas>=2.0.0
pykalman>=0.9.5
scikit-learn>=1.2.0
osmnx>=1.3.0
googlemaps>=4.10.0
fpdf>=1.7.2
Pillow>=9.5.0
pyyaml>=6.0.0